added experimental support for ironing and debug mode

+++ main.py
@@ -7,10 +7,11 @@ import os
 import transform as tf
 
 if __name__ == "__main__":
-    #orig_stl_path = 'test_files/test_pa_outline_fein_2.stl'
     orig_stl_path = 'test_files/test_slope_pa_2.stl'
-    prusa_config_path = 'test_files/generic_config_Deltiq2_lower_retraction.ini'
+    prusa_config_path = 'test_files/generic_config_Deltiq2_ironing.ini'
+    debug=True
     numPlanarBaseLayers = 2
+
     orig_stl = fr.openSTL(orig_stl_path)
     upscaled_stl = sf.upscale_stl(orig_stl, 3)
     filtered_surface, limits = sf.create_surface(upscaled_stl,np.deg2rad(80))
@@ -19,7 +20,7 @@ if __name__ == "__main__":
     print('Fall 2')
     ini_config = fr.slicer_config(fr.openINI(prusa_config_path))
     layer_height = ini_config.get_config_param('layer_height')
-    planarBaseOffset = numPlanarBaseLayers * layer_height
+    planarBaseOffset = numPlanarBaseLayers * float(layer_height)
 
     ps.sliceSTL(orig_stl_path,prusa_config_path,'--skirts 2 --skirt-height 2 --skirt-distance 6','C:\Program Files\Prusa3D\PrusaSlicer')
     planar_base_gcode, prusa_generated_config_planar = fr.openGCODE_keepcoms('output.gcode')
@@ -30,8 +31,12 @@ if __name__ == "__main__":
 
     ps.repairSTL(temp_stl_path)
     ps.sliceSTL(temp_stl_path,prusa_config_path,'','C:\Program Files\Prusa3D\PrusaSlicer')
-    os.remove(temp_stl_path)
     planar_gcode, prusa_generated_config = fr.openGCODE_keepcoms('output.gcode')
-    tf.transformGCODE(planar_gcode, base_layer_gcode, orig_stl_path, filtered_surface, prusa_generated_config)
+    tf.transformGCODE(planar_gcode, base_layer_gcode, orig_stl_path, planarBaseOffset,filtered_surface, prusa_generated_config, layer_height)
     
+    if debug==False:
+        os.remove(temp_stl_path)
+        os.remove('output.gcode')
+        os.remove('temp_gcode.gcode')
+
